Added test data - Protein-RNA(modified) complex
Complex of protein with RNA containing a non-standard base, N-MethylAdenosine (6MZ)

Sequence, dihedral and hbond restraints were taken from a DOCR
project with some manual modifications.
Because of problems in the DOCR project,
restraints were read from the original deposited Cyana restraint list
(as downloaded from the BMRB restraints grid) with custom
scripts and manual post-processing.

This is a good example of why NEF is necessary to avoid
the conversion problems encountered both in making the DOCR
project and in generating this test file.